+++ 3d/3d_camera.h
@@ -14,16 +14,16 @@ typedef struct _3DCamera
      *      y: -width/2 ~ width/2
      *      z: -height/2 ~ height/2
      */
-    uint32_t width;   //摄像屏幕宽
-    uint32_t height;  //摄像屏幕高
-    double ar;        //屏幕宽高比: ar = width / height
-    double openAngle; //相机视野开角,单位:度,范围[1,359]
-    uint32_t near;    //近端距离(屏幕到相机原点距离,即openAngle所在原点的距离)
-    uint32_t far;     //远端距离(远端到相机原点距离）
+    uint32_t width;  //摄像屏幕宽
+    uint32_t height; //摄像屏幕高
+    float ar;        //屏幕宽高比: ar = width / height
+    float openAngle; //相机视野开角,单位:度,范围[1,359]
+    uint32_t near;   //近端距离(屏幕到相机原点距离,即openAngle所在原点的距离)
+    uint32_t far;    //远端距离(远端到相机原点距离）
 
-    double xyz[3];      //相机原点当前所在坐标
-    double roll_xyz[3]; //相机绕自身坐标系转角(单位:度)
-    double lock_xyz[3]; //锁定目标点(就是让相机的旋转以此为原点)
+    float xyz[3];      //相机原点当前所在坐标
+    float roll_xyz[3]; //相机绕自身坐标系转角(单位:度)
+    float lock_xyz[3]; //锁定目标点(就是让相机的旋转以此为原点)
 
     uint32_t photoSize; //照片内存大小width*height*3
     uint8_t *photoMap;  //照片缓冲区,RGB存储格式,大小 width*height*3
@@ -48,11 +48,11 @@ typedef struct _3DCamera
 _3D_Camera *_3d_camera_init(
     uint32_t width,
     uint32_t height,
-    double openAngle,
+    float openAngle,
     uint32_t near,
     uint32_t far,
-    double *xyz,
-    double *roll_xyz);
+    float *xyz,
+    float *roll_xyz);
 
 // 相机重置
 void _3d_camera_reset(_3D_Camera *camera);
@@ -72,21 +72,21 @@ void _3d_camera_release(_3D_Camera **camera);
 /* ---------- 运动 ---------- */
 
 // 相机3轴旋转, 增量式, 绕自身坐标系, 单位:度
-void _3d_camera_roll(_3D_Camera *camera, double x, double y, double z);
+void _3d_camera_roll(_3D_Camera *camera, float x, float y, float z);
 
 // 相机3轴平移, 增量式, 基于空间坐标系
-void _3d_camera_mov(_3D_Camera *camera, double x, double y, double z);
+void _3d_camera_mov(_3D_Camera *camera, float x, float y, float z);
 
 // 相机3轴平移, 增量式, 基于自身坐标系
-void _3d_camera_mov2(_3D_Camera *camera, double x, double y, double z);
+void _3d_camera_mov2(_3D_Camera *camera, float x, float y, float z);
 
 /* ---------- 特效 ---------- */
 
 // 缩放, zoom为1时原始比例, 大于1放大图像, 小于1缩小图像
-void _3d_camera_zoom(_3D_Camera *camera, double zoom);
+void _3d_camera_zoom(_3D_Camera *camera, float zoom);
 
 // 锁定目标, 之后 _3d_camera_roll 将变成完全绕目标转动
-void _3d_camera_lock(_3D_Camera *camera, double *xyz);
+void _3d_camera_lock(_3D_Camera *camera, float *xyz);
 
 // 解除锁定
 void _3d_camera_unlock(_3D_Camera *camera);
